+++ tests/TestCase/Utility/StaticPageTest.php
@@ -54,9 +54,7 @@ class StaticPageTest extends TestCase
 
         $pages = StaticPage::all();
         $this->assertContainsOnlyInstancesOf(Entity::class, $pages);
-        foreach ($pages as $page) {
-            $this->assertInstanceOf(FrozenTime::class, $page->modified);
-        }
+        $this->assertContainsOnlyInstancesOf(FrozenTime::class, Hash::extract($pages, '{n}.modified'));
 
         //Checks filenames
         $this->assertEquals([
@@ -161,20 +159,17 @@ class StaticPageTest extends TestCase
      */
     public function testGetSlug()
     {
-        $getSlugMethod = function () {
-            return $this->invokeMethod(StaticPage::class, 'getSlug', func_get_args());
-        };
-
         foreach ([
-            'my-file', 'my-file.' . StaticPage::EXTENSION,
-            '/first/second/my-file.' . StaticPage::EXTENSION,
+            'my-file',
+            '/first/second/my-file',
         ] as $file) {
-            $this->assertEquals('my-file', $getSlugMethod($file, '/first/second'));
-            $this->assertEquals('my-file', $getSlugMethod($file, '/first/second/'));
+            $this->assertEquals('my-file', StaticPage::getSlug($file, '/first/second'));
+            $this->assertEquals('my-file', StaticPage::getSlug($file, '/first/second/'));
+            $this->assertEquals('my-file', StaticPage::getSlug($file . '.' . StaticPage::EXTENSION, '/first/second'));
         }
 
-        $this->assertEquals('first/my-file', $getSlugMethod('first/my-file.' . StaticPage::EXTENSION, '/first/second'));
-        $this->assertEquals('third/my-file', $getSlugMethod('/first/second/third/my-file.' . StaticPage::EXTENSION, '/first/second'));
+        $this->assertEquals('first/my-file', StaticPage::getSlug('first/my-file.' . StaticPage::EXTENSION, '/first/second'));
+        $this->assertEquals('third/my-file', StaticPage::getSlug('/first/second/third/my-file.' . StaticPage::EXTENSION, '/first/second'));
     }
 
     /**
@@ -184,15 +179,11 @@ class StaticPageTest extends TestCase
      */
     public function testGetSlugWin()
     {
-        $getSlugMethod = function () {
-            return $this->invokeMethod(StaticPage::class, 'getSlug', func_get_args());
-        };
-
-        $this->assertEquals('my-file', $getSlugMethod('\\first\\second\\my-file.' . StaticPage::EXTENSION, '\\first\\second'));
-        $this->assertEquals('my-file', $getSlugMethod('\\first\\second\\my-file.' . StaticPage::EXTENSION, '\\first\\second\\'));
-        $this->assertEquals('my-file', $getSlugMethod('C:\\\\first\\my-file.' . StaticPage::EXTENSION, 'C:\\\\first'));
-        $this->assertEquals('second/my-file', $getSlugMethod('\\first\\second\\my-file.' . StaticPage::EXTENSION, '\\first'));
-        $this->assertEquals('second/my-file', $getSlugMethod('\\first\\second\\my-file.' . StaticPage::EXTENSION, '\\first\\'));
+        $this->assertEquals('my-file', StaticPage::getSlug('\\first\\second\\my-file.' . StaticPage::EXTENSION, '\\first\\second'));
+        $this->assertEquals('my-file', StaticPage::getSlug('\\first\\second\\my-file.' . StaticPage::EXTENSION, '\\first\\second\\'));
+        $this->assertEquals('my-file', StaticPage::getSlug('C:\\\\first\\my-file.' . StaticPage::EXTENSION, 'C:\\\\first'));
+        $this->assertEquals('second/my-file', StaticPage::getSlug('\\first\\second\\my-file.' . StaticPage::EXTENSION, '\\first'));
+        $this->assertEquals('second/my-file', StaticPage::getSlug('\\first\\second\\my-file.' . StaticPage::EXTENSION, '\\first\\'));
     }
 
     /**